requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
idna==3.10
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
